Created Python_Image_Processing
Looks like there's only an input and then 4 functional lines of code. Hopefully this is sufficient for the whole project, but we'll see.

This code prints out the entire array of Black and White pixel data, and then prints out a 50x50 2D array of that data. 
0 is totally black, 255 is totally white.
It then shows an upscaled output and an actual size output for verification.